+++ app/src/main/java/com/battlelancer/seriesguide/SgApp.kt
@@ -93,6 +93,8 @@ class SgApp : Application() {
 
         const val RELEASE_VERSION_2025_2_6 = 21250207
 
+        const val RELEASE_VERSION_2026_2_0 = 23260200
+
         /**
          * The content authority used to identify the SeriesGuide [android.content.ContentProvider].
          */