+++ app/src/main/java/com/battlelancer/seriesguide/provider/SeriesGuideContract.java
@@ -10,12 +10,13 @@ import androidx.annotation.IntDef;
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
 import com.battlelancer.seriesguide.SgApp;
+import com.battlelancer.seriesguide.movies.database.SgMovie;
 import com.battlelancer.seriesguide.shows.database.SgEpisode2;
 import com.battlelancer.seriesguide.shows.database.SgShow2;
 import com.battlelancer.seriesguide.shows.episodes.EpisodeFlags;
+import com.battlelancer.seriesguide.shows.tools.NextEpisodeUpdater;
 import com.battlelancer.seriesguide.shows.tools.ShowStatus;
 import com.battlelancer.seriesguide.shows.tools.ShowTools2;
-import com.battlelancer.seriesguide.shows.tools.NextEpisodeUpdater;
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
 
@@ -488,7 +489,7 @@ public class SeriesGuideContract {
          */
         String POSTER = "series_poster";
         /**
-         * Path to a small variant of the poster. Needs to be prefixed with the poster server URL.
+         * See {@link SgShow2#getPosterSmall()}.
          */
         String POSTER_SMALL = "series_poster_small";
 
@@ -499,6 +500,9 @@ public class SeriesGuideContract {
          */
         String STATUS = "series_status";
 
+        /**
+         * See {@link SgShow2#getRuntime()}.
+         */
         String RUNTIME = "series_runtime";
 
         /** See {@link SgShow2#getRatingTmdb()}. */
@@ -532,12 +536,7 @@ public class SeriesGuideContract {
         String FIRST_RELEASE = "series_firstaired";
 
         /**
-         * Local release time. Encoded as integer (hhmm).
-         *
-         * <pre>
-         * Example: 2035
-         * Default: -1
-         * </pre>
+         * See {@link SgShow2#getReleaseTime()}.
          */
         String RELEASE_TIME = "series_airstime";
 
@@ -941,7 +940,8 @@ public class SeriesGuideContract {
             ListItemTypes.TVDB_SHOW,
             ListItemTypes.SEASON,
             ListItemTypes.EPISODE,
-            ListItemTypes.TMDB_SHOW
+            ListItemTypes.TMDB_SHOW,
+            ListItemTypes.TMDB_MOVIE
     })
     public @interface ListItemTypes {
         int TVDB_SHOW = 1;
@@ -954,9 +954,10 @@ public class SeriesGuideContract {
          */
         int EPISODE = 3;
         int TMDB_SHOW = 4;
+        int TMDB_MOVIE = 5;
     }
 
-    interface MoviesColumns {
+    public interface MoviesColumns {
 
         String TITLE = "movies_title";
 
@@ -969,6 +970,9 @@ public class SeriesGuideContract {
 
         String TMDB_ID = "movies_tmdbid";
 
+        /**
+         * See {@link SgMovie#getPoster()}.
+         */
         String POSTER = "movies_poster";
 
         String GENRES = "movies_genres";
@@ -976,11 +980,13 @@ public class SeriesGuideContract {
         String OVERVIEW = "movies_overview";
 
         /**
-         * Release date in milliseconds. Store Long.MAX if unknown, as it is likely in the future
-         * (also helps correctly sorting movies by release date).
+         * See {@link SgMovie#getReleasedMsOrDefault()}.
          */
         String RELEASED_UTC_MS = "movies_released";
 
+        /**
+         * See {@link SgMovie#getRuntimeMin()}.
+         */
         String RUNTIME_MIN = "movies_runtime";
 
         String TRAILER = "movies_trailer";
@@ -1062,8 +1068,6 @@ public class SeriesGuideContract {
 
     public static final String PATH_LIST_ITEMS = "listitems";
 
-    public static final String PATH_WITH_DETAILS = "with_details";
-
     public static final String PATH_MOVIES = "movies";
 
     public static final String PATH_JOBS = "jobs";
@@ -1136,14 +1140,6 @@ public class SeriesGuideContract {
                 .build();
 
         /**
-         * List items table joined with shows, seasons and episodes table (depending on list item
-         * type). See {@link SeriesGuideProvider#LIST_ITEMS_WITH_DETAILS}.
-         */
-        public static final Uri CONTENT_WITH_DETAILS_URI = CONTENT_URI.buildUpon()
-                .appendPath(PATH_WITH_DETAILS)
-                .build();
-
-        /**
          * Use if multiple items get returned
          */
         public static final String CONTENT_TYPE = "vnd.android.cursor.dir/vnd.seriesguide.listitem";
@@ -1155,16 +1151,6 @@ public class SeriesGuideContract {
                 = "vnd.android.cursor.item/vnd.seriesguide.listitem";
 
         public static final String SELECTION_LIST = Lists.LIST_ID + "=?";
-        public static final String SELECTION_TVDB_SHOWS =
-                ListItems.TYPE + "=" + ListItemTypes.TVDB_SHOW;
-        public static final String SELECTION_TMDB_SHOWS =
-                ListItems.TYPE + "=" + ListItemTypes.TMDB_SHOW;
-        public static final String SELECTION_SEASONS =
-                ListItems.TYPE + "=" + ListItemTypes.SEASON;
-        public static final String SELECTION_EPISODES =
-                ListItems.TYPE + "=" + ListItemTypes.EPISODE;
-
-        public static final String SORT_TYPE = ListItems.TYPE + " ASC";
 
         public static Uri buildListItemUri(String id) {
             return CONTENT_URI.buildUpon().appendPath(id).build();
@@ -1202,7 +1188,8 @@ public class SeriesGuideContract {
             return type == ListItemTypes.TVDB_SHOW
                     || type == ListItemTypes.SEASON
                     || type == ListItemTypes.EPISODE
-                    || type == ListItemTypes.TMDB_SHOW;
+                    || type == ListItemTypes.TMDB_SHOW
+                    || type == ListItemTypes.TMDB_MOVIE;
         }
     }
 
@@ -1219,14 +1206,9 @@ public class SeriesGuideContract {
 
         public static final String SELECTION_COLLECTION = Movies.IN_COLLECTION + "=1";
 
-        public static final String SELECTION_NOT_COLLECTION = Movies.IN_COLLECTION + "=0";
-
         public static final String SELECTION_WATCHLIST = Movies.IN_WATCHLIST + "=1";
 
-        public static final String SELECTION_NOT_WATCHLIST = Movies.IN_WATCHLIST + "=0";
-
         public static final String SELECTION_WATCHED = Movies.WATCHED + "=1";
-        public static final String SELECTION_UNWATCHED = Movies.WATCHED + "=0";
 
         // Android provides the UNICODE collator,
         // use to correctly order characters with e.g. accents.